surchargeDTO, imageDataDTO created, get and post Surcharge interfaces renamed
