+++ functions/src/data/surcharges/DTO/imageDataDTO.ts
@@ -1,0 +1,6 @@
+export type ImageDataDTO = {
+    base64: string; // Base64 representation of the image
+    width: number;  // Image width
+    height: number; // Image height
+    format: string; // e.g., "png", "jpg"
+  };
