+++ functions/src/data/surcharges/DTO/surchargeDTO.ts
@@ -1,0 +1,11 @@
+import {ImageDataDTO} from "./imageDataDTO"
+
+export type SurchargeDTO = {
+    id: string,
+    rate: number,
+    totalAmount?: number,
+    surchargeAmount?: number,
+    purchaseAmount?: number,
+    imageData?: ImageDataDTO[]
+}
+
